Fix QUIET M28 Factory Building
This Commit fixes a large issue on about 4 maps I tested and played and reviewed replays of where M28 would not build enough factories due to have too strict of Energy Requirements. This fixes that and also solves the problem of aggressively teching too early. These fixes are QUIET Specific as well so they do NOT affect FAF, LOUD, or Steam.